Change data type of poolValue to int

+++ src/main/java/org/terasology/PointPool/PointPoolCommands.java
@@ -37,7 +37,7 @@ public class PointPoolCommands extends BaseComponentSystem {
 
     @Command(value = "fillPool", shortDescription = "Fill pool by amount given",
             requiredPermission = PermissionManager.NO_PERMISSION)
-    public String fill(@Sender EntityRef client, @CommandParam("amount") float amount) {
+    public String fill(@Sender EntityRef client, @CommandParam("amount") int amount) {
         ClientComponent clientComp = client.getComponent(ClientComponent.class);
 
         clientComp.character.send(new FillPoolEvent(amount, clientComp.character));
@@ -47,7 +47,7 @@ public class PointPoolCommands extends BaseComponentSystem {
 
     @Command(value = "drainPool", shortDescription = "Drain pool by amount given",
             requiredPermission = PermissionManager.NO_PERMISSION)
-    public String drainPool(@Sender EntityRef client, @CommandParam("amount") float amount) {
+    public String drainPool(@Sender EntityRef client, @CommandParam("amount") int amount) {
         ClientComponent clientComp = client.getComponent(ClientComponent.class);
         clientComp.character.send(new DrainPoolEvent(amount, clientComp.character));
         return "Pool drained by " + amount;
@@ -86,7 +86,7 @@ public class PointPoolCommands extends BaseComponentSystem {
         // TODO : Make this work for multiple pools attached to same entity
         // for(PointPoolComponent component : clientComp.character.getComponent(PointPoolComponent.class)
         try {
-            float value = clientComp.character.getComponent(PointPoolComponent.class).poolValue;
+            int value = clientComp.character.getComponent(PointPoolComponent.class).poolValue;
             return "Current pool value is " + value;
         } catch (NullPointerException n) {
             return "NullPointerException encountered";
